fix: move agents principal helper out of runtime
