+++ .changeset/clean-agents-report.md
@@ -1,6 +1,5 @@
 ---
 "@electric-ax/agents": patch
-"@electric-ax/agents-runtime": patch
 "electric-ax": patch
 ---
