+++ packages/agents-runtime/src/runtime-server-client.ts
@@ -1,7 +1,8 @@
 import type { EntityTags, TagOperation } from './tags'
-import { ELECTRIC_PRINCIPAL_HEADER } from './headers'
 import { appendPathToUrl } from './url'
 import type { ClaimTokenHeader, HeadersProvider } from './types'
+
+const ELECTRIC_PRINCIPAL_HEADER = `electric-principal`
 
 export interface RuntimeServerClientConfig {
   baseUrl: string
