+++ packages/agents/src/server-headers.ts
