+++ packages/agents/tsdown.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'tsdown'
 
 export default defineConfig([
   {
-    entry: [`src/index.ts`],
+    entry: [`src/index.ts`, `src/server-headers.ts`],
     format: [`esm`, `cjs`],
     platform: `node`,
     dts: true,
